Prefer visibilitychange over focus and blur
The ClassCycler would pause on blur, even if the window was still visible but unfocused. The ClassCycler now attaches a visibilitychange event listener on the window, if the browser permits. Otherwise, it resorts to the original focus and blur event listeners.

This change also reverts the handler registry back to an object whose keys are the event types. This allows greater flexibility of working with the handler registry because the mapping of event listeners to its associated event handlers during removal is not as dependent on the registry's order of stored event handlers.